+++ src/client/session.hpp
@@ -24,19 +24,19 @@ public:
 
   // Get number of data in this session (determined by # of data in DB and
   // SessionOption).
-  BigInt GetNumData() const;
+  int64_t GetNumData() const;
 
   // Create an iterator that sequentially returns data [data_begin, data_end).
   // Default to include all data.
   /*
-  DataIterator NewDataIterator(BigInt data_begin = 0,
-      BigInt data_end = -1) const;
+  DataIterator NewDataIterator(size_t data_begin = 0,
+      size_t data_end = -1) const;
   */
 
-  DataIterator NewDataIterator(BigInt data_begin = 0,
-        BigInt data_end = -1, bool use_multi_threads = true,
-      BigInt num_io_threads = 1, BigInt num_transform_threads = 4,
-      BigInt buffer_limit = 16, BigInt batch_limit = 16) const;
+  DataIterator NewDataIterator(size_t data_begin = 0,
+        size_t data_end = -1, bool use_multi_threads = true,
+      int32_t num_io_threads = 1, int32_t num_transform_threads = 4,
+      size_t buffer_limit = 16, size_t batch_limit = 16) const;
 
   Status GetStatus() const;
